requirements/base.txt
# Base requirements
pandas>=1.3.0
numpy>=1.20.0
requests>=2.25.0
praw>=7.5.0
pyyaml>=6.0
scikit-learn>=1.0.0
nltk>=3.6.0
transformers>=4.15.0
sqlalchemy>=1.4.0
python-dotenv>=0.19.0
